Add a not_found() helper for typical usage of NotFound
Some direct usage exists so the NotFound type is also re-exported.
