+++ src/router.rs
@@ -178,7 +178,7 @@ impl Handler for R404 {
                 req.mut_extensions().insert(m.params.clone());
                 m.handler.call(req)
             }
-            Err(..) => Ok(NotFound.response().unwrap()),
+            Err(..) => Ok(NotFound.into()),
         }
     }
 }
@@ -187,7 +187,7 @@ impl Handler for R404 {
 mod tests {
     use super::*;
     use crate::util::errors::{
-        bad_request, cargo_err, forbidden, internal, AppError, ChainError, NotFound,
+        bad_request, cargo_err, forbidden, internal, not_found, AppError, ChainError,
     };
     use crate::util::EndpointResult;
 
@@ -220,7 +220,7 @@ mod tests {
             StatusCode::NOT_FOUND
         );
         assert_eq!(
-            C(|_| err(NotFound)).call(&mut req).unwrap().status(),
+            C(|_| Err(not_found())).call(&mut req).unwrap().status(),
             StatusCode::NOT_FOUND
         );
